django_requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# AI & Computer Vision (동일)
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
transformers==4.35.2
accelerate==0.24.1
httpx==0.25.2
scikit-image==0.21.0

# Django Extensions
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.26.5  # API 문서화 (Swagger)

# Production
gunicorn==21.2.0
whitenoise==6.6.0